Revert "1563"
This reverts commit 71c963d9a9918a78421b8849fe0a25c65350b678.

+++ src/playground_base/playground_base/gui/glue/FBHostGUIContext.cpp
@@ -1,7 +1,7 @@
 #include <playground_base/gui/glue/FBHostGUIContext.hpp>
 
 void 
-FBHostGUIContext::PerformImmediateAudioParamEdit(int index, double normalized)
+FBHostGUIContext::PerformImmediateAudioParamEdit(int index, float normalized)
 {
   BeginAudioParamChange(index);
   PerformAudioParamEdit(index, normalized);
